Remove dependency on pyscopg2
The python package pyscopg2 is a large requirement that also brings in 
heavy dependencies (postgresql-server-dev-*). However, in dmrunner it is 
only used to check whether the postgresql server is running.

This commit changes the psql healthcheck to use the pg_admin tool that 
is included with postgresql-client. While this is still a fairly large 
dependency, it is lesser than requiring an entire psql server that 
should not even be running.